+++ ddm/participation/urls.py
@@ -8,6 +8,11 @@ urlpatterns = [
     path(r"briefing/", views.BriefingView.as_view(), name="briefing"),
     path(r"data-donation/", views.DataDonationView.as_view(), name="datadonation"),
     path(r"questionnaire/", views.QuestionnaireView.as_view(), name="questionnaire"),
+    path(
+        r"questionnaire/progress/",
+        views.QuestionnaireProgressView.as_view(),
+        name="questionnaire_progress",
+    ),
     path(r"debriefing/", views.DebriefingView.as_view(), name="debriefing"),
     path(r"continue/", views.ContinuationView.as_view(), name="continuation"),
     path(r"inactive/", views.ProjectInactiveView.as_view(), name="project_inactive"),